# Improve Iter() performance.