Channels are much slower than calling a function

```
1m results:
pkg: github.com/GoogleCloudPlatform/testgrid/internal/result
BenchmarkIterSlow-10            78166888                14.90 ns/op
BenchmarkIterFast-10            759703227                1.559 ns/op

10m results:
pkg: github.com/GoogleCloudPlatform/testgrid/internal/result
BenchmarkIterSlow-10                   1        1360214667 ns/op
BenchmarkIterFast-10            728631373                1.590 ns/op
```